Data handling&logging (#7)
* Added Logger Implementation Which writes logs
with writeLog method and errors with writeError Method

* note adde,d please read it

* Making a CSV file handling implementation.

- Editted the editRecord in the Attendance Data
Handler interface to take a list of attendance records
instade of a single record. And made some edits in the docs.

- Added AttendanceFileHandler implementation of the
AttendanceDataHandler. Which stores the data in a CSV file
in the folder "./data". Introduced some helper methods to make the
code clearer to read and maintain.

* Putting the final touches for the first virsion.

- Renamed LoggerImpl to Logger. Made its attributes private and final.
- Created the /data floder to store the attendance recorrds data.
- Made AttendanceFileHandler's attributes private and final

---------
Authored-by Amr
Co-authored-by: Luai Waleed